pyproject.toml
[project]
name = "try-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.6.0",
    "openai>=1.70.0",
    "python-dotenv>=1.1.0",
]
